+++ src/main/java/uk/gov/hmcts/reform/rd/commondata/camel/util/CommonDataLoadConstants.java
@@ -7,6 +7,8 @@ public final class CommonDataLoadConstants {
     }
 
     public static final String IS_READY_TO_AUDIT = "IS_READY_TO_AUDIT";
+
+    public static final String FILE_NAME = "file-name";
     public static final String FLAG_CODE = "flag_code";
     public static final String FLAG_CODE_NOT_EXISTS = "flag_code does not exist in parent table";
     public static final String ROUTER_NAME = "RD_Common_Data";
